Use python 3.5 and get rid of warnings
