+++ saleor/order/utils.py
@@ -1,0 +1,47 @@
+import logging
+
+from payments.signals import status_changed
+from django.dispatch import receiver
+from django.shortcuts import get_object_or_404, redirect
+from functools import wraps
+
+from .models import Order
+from ..core import analytics
+
+
+logger = logging.getLogger(__name__)
+
+
+def check_order_status(func):
+    @wraps(func)
+    def decorator(*args, **kwargs):
+        token = kwargs.pop('token')
+        order = get_object_or_404(Order, token=token)
+        if order.is_fully_paid():
+            return redirect('order:details', token=order.token)
+        kwargs['order'] = order
+        return func(*args, **kwargs)
+
+    return decorator
+
+
+@receiver(status_changed)
+def order_status_change(sender, instance, **kwargs):
+    order = instance.order
+    if order.is_fully_paid():
+        order.change_status('fully-paid')
+        instance.send_confirmation_email()
+        try:
+            analytics.report_order(order.tracking_client_id, order)
+        except Exception:
+            # Analytics failing should not abort the checkout flow
+            logger.exception('Recording order in analytics failed')
+
+
+def get_ip(request):
+    ip = request.META.get("HTTP_X_FORWARDED_FOR", None)
+    if ip:
+        ip = ip.split(", ")[0]
+    else:
+        ip = request.META.get("REMOTE_ADDR", "")
+    return ip
