+++ tox.ini
@@ -1,9 +1,10 @@
 [tox]
-envlist = py{27,33,34}-django{18,_master}
+envlist = py{27,34,35}-django{18,19,_master}
 
 [testenv]
 deps =
-    django18: django>=1.8,<1.9
+    django18: django>=1.8,<1.9a0
+    django19: django>=1.9a1,<1.10a0
     django_master: https://github.com/django/django/archive/master.tar.gz
     mock==1.3.0
     purl>=0.4.1
